ödev ilavesi

+++ Kdv.java
@@ -7,7 +7,7 @@ public class Kdv {
         Scanner inp = new Scanner(System.in);
 
         System.out.println("Bir tutar giriniz ?");
-        deger=inp.nextFloat();
+        deger=inp.nextDouble();
         inp.close();
         if(deger <=1000){kdvorani= 0.18;}
         else{kdvorani=0.08;}
